fix clang build issues

+++ include/intx/intx.hpp
@@ -164,7 +164,7 @@ struct result_with_carry
 /// @{
 
 /// Addition with carry. `uint64_t *carry` is used as in/out parameter
-inline constexpr uint64_t addc(uint64_t x, uint64_t y, uint64_t *carry) noexcept
+inline constexpr uint64_t addc(uint64_t x, uint64_t y, unsigned long long *carry) noexcept
 {
 #if __has_builtin(__builtin_addcll)
     if (!is_constant_evaluated())
@@ -190,7 +190,7 @@ inline constexpr uint64_t addc(uint64_t x, uint64_t y, uint64_t *carry) noexcept
 }
 
 /// Subtraction with carry (borrow). `uint64_t *carry` is used as in/out parameter
-inline constexpr uint64_t subc(uint64_t x, uint64_t y, uint64_t *carry) noexcept
+inline constexpr uint64_t subc(uint64_t x, uint64_t y, unsigned long long *carry) noexcept
 {
 #if __has_builtin(__builtin_subcll)
     if (!is_constant_evaluated())
@@ -217,7 +217,7 @@ inline constexpr uint64_t subc(uint64_t x, uint64_t y, uint64_t *carry) noexcept
 
 /// Addition with carry.
 template <unsigned N>
-inline constexpr uint<N> addc(const uint<N>& x, const uint<N>& y, uint64_t *carry) noexcept
+inline constexpr uint<N> addc(const uint<N>& x, const uint<N>& y, unsigned long long *carry) noexcept
 {
     uint<N> s;
     for (size_t i = 0; i < uint<N>::num_words; ++i)
@@ -229,7 +229,7 @@ inline constexpr uint<N> addc(const uint<N>& x, const uint<N>& y, uint64_t *carr
 
 inline constexpr uint128 operator+(uint128 x, uint128 y) noexcept
 {
-    uint64_t carry = 0;
+    unsigned long long carry = 0;
     return addc(x, y, &carry);
 }
 
@@ -241,7 +241,7 @@ inline constexpr uint128 operator+(uint128 x) noexcept
 /// Performs subtraction of two unsigned numbers and returns the difference
 /// and the carry bit (aka borrow, overflow).
 template <unsigned N>
-inline constexpr uint<N> subc(const uint<N>& x, const uint<N>& y, uint64_t *carry) noexcept
+inline constexpr uint<N> subc(const uint<N>& x, const uint<N>& y, unsigned long long *carry) noexcept
 {
     uint<N> z;
     for (size_t i = 0; i < uint<N>::num_words; ++i)
@@ -253,7 +253,7 @@ inline constexpr uint<N> subc(const uint<N>& x, const uint<N>& y, uint64_t *carr
 
 inline constexpr uint128 operator-(uint128 x, uint128 y) noexcept
 {
-    uint64_t carry = 0;
+    unsigned long long carry = 0;
     return subc(x, y, &carry);
 }
 
@@ -1125,7 +1125,7 @@ inline constexpr bool operator<(const uint256& x, const uint256& y) noexcept
 template <unsigned N>
 inline constexpr bool operator<(const uint<N>& x, const uint<N>& y) noexcept
 {
-    uint64_t carry = 0;
+    unsigned long long carry = 0;
     subc(x, y, &carry);
     return !!carry;
 }
@@ -1299,7 +1299,7 @@ inline constexpr uint<N> operator<<(const uint<N>& x, uint64_t shift) noexcept
     const auto skip = static_cast<size_t>(shift / word_bits);
 
     uint<N> r;
-    uint64_t carry = 0;
+    unsigned long long carry = 0;
     for (size_t i = 0; i < (uint<N>::num_words - skip); ++i)
     {
         r[i + skip] = (x[i] << s) | carry;
@@ -1352,7 +1352,7 @@ inline constexpr uint<N> operator>>(const uint<N>& x, uint64_t shift) noexcept
     const auto skip = static_cast<size_t>(shift / word_bits);
 
     uint<N> r;
-    uint64_t carry = 0;
+    unsigned long long carry = 0;
     for (size_t i = 0; i < (num_words - skip); ++i)
     {
         r[num_words - 1 - i - skip] = (x[num_words - 1 - i] >> s) | carry;
@@ -1451,7 +1451,7 @@ inline const uint8_t* as_bytes(const T& x) noexcept
 template <unsigned N>
 inline constexpr uint<N> operator+(const uint<N>& x, const uint<N>& y) noexcept
 {
-    uint64_t carry = 0;
+    unsigned long long carry = 0;
     return addc(x, y, &carry);
 }
 
@@ -1464,7 +1464,7 @@ inline constexpr uint<N> operator-(const uint<N>& x) noexcept
 template <unsigned N>
 inline constexpr uint<N> operator-(const uint<N>& x, const uint<N>& y) noexcept
 {
-    uint64_t carry = 0;
+    unsigned long long carry = 0;
     return subc(x, y, &carry);
 }
 
@@ -1493,7 +1493,7 @@ inline constexpr uint<2 * N> umul(const uint<N>& x, const uint<N>& y) noexcept
         uint64_t k = 0;
         for (size_t i = 0; i < num_words; ++i)
         {
-            uint64_t carry = 0;
+            unsigned long long carry = 0;
             const auto a = addc(p[i + j], k, &carry);
             const auto t = umul(x[i], y[j]) + uint128{a, carry};
             p[i + j] = t[0];
@@ -1517,7 +1517,7 @@ inline constexpr uint<N> operator*(const uint<N>& x, const uint<N>& y) noexcept
         uint64_t k = 0;
         for (size_t i = 0; i < (num_words - j - 1); i++)
         {
-            uint64_t carry = 0;
+            unsigned long long carry = 0;
             const auto a = addc(p[i + j], k, &carry);
             const auto t = umul(x[i], y[j]) + uint128{a, carry};
             p[i + j] = t[0];
@@ -1709,7 +1709,7 @@ inline bool add(uint64_t s[], const uint64_t x[], const uint64_t y[], int len) n
     // OPT: Add MinLen template parameter and unroll first loop iterations.
     INTX_REQUIRE(len >= 2);
 
-    uint64_t carry = 0;
+    unsigned long long carry = 0;
     for (int i = 0; i < len; ++i)
         s[i] = addc(x[i], y[i], &carry);
     return !!carry;
@@ -1761,9 +1761,9 @@ inline void udivrem_knuth(
             std::tie(qhat, rhat) = udivrem_3by2(u2, u1, u0, divisor, reciprocal);
 
             const auto overflow = submul(&u[j], &u[j], d, dlen - 2, qhat);
-            uint64_t carry1 = 0;
+            unsigned long long carry1 = 0;
             u[j + dlen - 2] = subc(rhat[0], overflow, &carry1);
-            uint64_t carry2 = 0;
+            unsigned long long carry2 = 0;
             u[j + dlen - 1] = subc(rhat[1], carry1, &carry2);
 
             if (INTX_UNLIKELY(!!carry2))
@@ -2030,7 +2030,7 @@ inline uint256 addmod(const uint256& x, const uint256& y, const uint256& mod) no
     {
         // Normalize x in case it is bigger than mod.
         auto xn = x;
-        uint64_t carry = 0;
+        unsigned long long carry = 0;
         const auto xd = subc(x, mod, &carry);
         if (!carry)
             xn = xd;
@@ -2043,7 +2043,7 @@ inline uint256 addmod(const uint256& x, const uint256& y, const uint256& mod) no
             yn = yd;
 
         carry = 0;
-        uint64_t carry2 = 0;
+        unsigned long long carry2 = 0;
         const auto av = addc(xn, yn, &carry2);
         const auto bv = subc(av, mod, &carry);
         if (carry2 || !carry)
@@ -2051,7 +2051,7 @@ inline uint256 addmod(const uint256& x, const uint256& y, const uint256& mod) no
         return av;
     }
 
-    uint64_t carry = 0;
+    unsigned long long carry = 0;
     const auto s = addc(x, y, &carry);
     uint<256 + 64> n = s;
     n[4] = carry;
